// src/shared/api/generated/models/cmsUserDetailDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Lola Backend Platform
 * CMS, integration, chat and realtime API for Lola AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import type { CmsUserDetailDtoStatus } from "./cmsUserDetailDtoStatus";

export interface CmsUserDetailDto {
  createdAt: string;
  /** @nullable */
  deactivatedAt: string | null;
  /** @nullable */
  deactivationReason: string | null;
  displayName: string;
  email: string;
  emailVerified: boolean;
  familyName: string;
  givenName: string;
  id: string;
  /** @nullable */
  lastLoginAt: string | null;
  platformPermissionCodes: string[];
  platformRoleKeys: string[];
  projectCount: number;
  status: CmsUserDetailDtoStatus;
  updatedAt: string;
  version: number;
}