// src/shared/api/generated/models/cmsUserLifecycleListParams.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Lola Backend Platform
 * CMS, integration, chat and realtime API for Lola AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import type { CmsUserLifecycleListStatus } from "./cmsUserLifecycleListStatus";

export type CmsUserLifecycleListParams = {
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: number;
  /**
   * @minLength 1
   * @maxLength 512
   */
  cursor?: string;
  status?: CmsUserLifecycleListStatus;
};